+++ example/ext/sqla/model/project.py
@@ -1,0 +1,5 @@
+from . import Base
+
+
+class Project(Base):
+    __tablename__ = 'project'