+++ CHANGES.rst
@@ -3,6 +3,8 @@ Changelog
 
 Development
 -----------
+- Add sqlalchemy extension.
+- Change the way links are added to resource representation.
 
 - Fix issue #6: HTTP 500 error when using POST verb on Item resources.
 - Breaking changes: